All possible paths
An attempt to learn Python and Git - A solution to a fun puzzle I dug
up: A robot walks from the top-left to the bottom-right corner of an NxN
grid without visiting the same spot twice. How many possible paths are
there if the robot moves square by square (non-diagonally)?